Moved functions between files in compiler module
