+++ compiler/program.py
@@ -1,8 +1,9 @@
 import argparse
 import subprocess
 from typing import List
+from compiler.asm import initialize_asm, generate_asm
+from compiler.compile import compile_asm, link_object_file
 from compiler.defs import TokenType, Token, OpType, Program, Op, Intrinsic
-from compiler.asm import initialize_asm, generate_asm, compile_asm, link_object_file
 
 def intrinsic_exists(token: str) -> bool:
     return bool(hasattr(Intrinsic, token))
